sigclip_iters should be int rather than float

+++ photutils/detection/core.py
@@ -68,7 +68,7 @@ def detect_threshold(data, snr, background=None, error=None, mask=None,
         The number of standard deviations to use as the clipping limit
         when calculating the image background statistics.
 
-    sigclip_iters : float, optional
+    sigclip_iters : int, optional
        The number of iterations to perform sigma clipping, or `None` to
        clip until convergence is achieved (i.e., continue until the last
        iteration clips nothing) when calculating the image background
